Create 1stOct_Vasav-Y.cpp
I've included iostream and cmath header files. Main function takes from user the input n and then calls the function fib which takes an integer as input. The use of fib function is to print the Fibonacci series with a twist, in order to do so before printing any of the value this function calls the isPrime function. Which takes an integer as an input and returns a bool value as an output. If the integer given to this function is an integer it will return true and if the integer entered is 1 or not prime it'll return false. And then we finally check whether the nth number is divisible by 5 or not.